requirements.txt
PyJWT
flask
numpy
joblib
sqlalchemy
scikit-learn